Fix JUnit message / type fields.
The message and type attributes never worked because they were expected to be nodes rather than attributes.

Note: in some cases it might have previously looked like it worked because when 'message' is not provided the annotation defaults to the first line of 'details' which in some cases happens to be the error.